requirements.txt
beautifulsoup4
requests
openai
arxiv
python-dotenv